# docs/swagger.yaml
basePath: /
definitions:
  db.TableResponse:
    properties:
      data:
        items:
          additionalProperties: true
          type: object
        type: array
      meta:
        $ref: '#/definitions/db.tableResponseMeta'
    type: object
  db.property:
    properties:
      desc:
        type: boolean
      name:
        type: string
      remap:
        type: string
      table:
        type: string
    type: object
  db.tableResponseMeta:
    properties:
      available_props:
        items:
          $ref: '#/definitions/db.property'
        type: array
      count:
        type: integer
      included_props:
        items:
          $ref: '#/definitions/db.property'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  tables.ServiceTag:
    properties:
      created:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      svc_id:
        type: string
      tag_attach_data:
        items:
          type: integer
        type: array
      tag_id:
        type: string
      updated_at:
        type: string
    type: object
  tables.Tag:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      tag_created:
        type: string
      tag_data:
        type: string
      tag_exclude:
        type: string
      tag_id:
        type: string
      tag_name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: collector-api-contact@opensvc.com
    name: OpenSVC SAS
    url: https://www.opensvc.com
  description: Organization clusters, nodes, services and more.
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: OpenSVC collector API
  version: "1.0"
paths:
...
  /services/{id}/candidate_tags:
    get:
      consumes:
      - application/json
      parameters:
      - description: properties to include, and optionally remap (comma separated)
        in: query
        name: props
        type: string
      - description: properties to group by (comma separated)
        in: query
        name: groupby
        type: string
      - description: properties to order by (comma separated, prefix with '~' to reverse)
        in: query
        name: order
        type: string
      - description: property value filter (a, !a, a&b, a|b, (a,b),  a%,  a%&!ab%)
        in: query
        items:
          type: string
        name: filters
        type: array
      - description: number of objets to include in response
        in: query
        name: limit
        type: integer
      - description: offset of the first objet to include in response
        in: query
        name: offset
        type: integer
      - description: turn off metadata in response
        in: query
        name: meta
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TableResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List existing tags not already attached to a service
      tags:
      - tags
  /services/{id}/tags:
    get:
      consumes:
      - application/json
      parameters:
      - description: properties to include, and optionally remap (comma separated)
        in: query
        name: props
        type: string
      - description: properties to group by (comma separated)
        in: query
        name: groupby
        type: string
      - description: properties to order by (comma separated, prefix with '~' to reverse)
        in: query
        name: order
        type: string
      - description: property value filter (a, !a, a&b, a|b, (a,b),  a%,  a%&!ab%)
        in: query
        items:
          type: string
        name: filters
        type: array
      - description: number of objets to include in response
        in: query
        name: limit
        type: integer
      - description: offset of the first objet to include in response
        in: query
        name: offset
        type: integer
      - description: turn off metadata in response
        in: query
        name: meta
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TableResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List tags attached to a service
      tags:
      - tags
  /services/{id}/tags/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: the index of the entry in database, or uuid, or name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tables.ServiceTag'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Show a tag attachment to a service
      tags:
      - services
  /services/tags:
    get:
      consumes:
      - application/json
      parameters:
      - description: properties to include, and optionally remap (comma separated)
        in: query
        name: props
        type: string
      - description: properties to group by (comma separated)
        in: query
        name: groupby
        type: string
      - description: properties to order by (comma separated, prefix with '~' to reverse)
        in: query
        name: order
        type: string
      - description: property value filter (a, !a, a&b, a|b, (a,b),  a%,  a%&!ab%)
        in: query
        items:
          type: string
        name: filters
        type: array
      - description: number of objets to include in response
        in: query
        name: limit
        type: integer
      - description: offset of the first objet to include in response
        in: query
        name: offset
        type: integer
      - description: turn off metadata in response
        in: query
        name: meta
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TableResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List tags attachments to services
      tags:
      - tags
  /services/tags/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: the index of the entry in database, or uuid, or name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tables.ServiceTag'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Show a tag attachment to a service
      tags:
      - services
  /tags:
    delete:
      consumes:
      - application/json
      description: Deleting tags requires the TagManager privilege.
      parameters:
      - description: property value filter (a, !a, a&b, a|b, (a,b),  a%,  a%&!ab%)
        in: query
        items:
          type: string
        name: filters
        type: array
      - description: properties to order by (comma separated, prefix with '~' to reverse)
        in: query
        name: order
        type: string
      - description: number of objets to include in response
        in: query
        name: limit
        type: integer
      - description: offset of the first objet to include in response
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tables.Tag'
            type: array
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Delete tags
      tags:
      - tags
    get:
      consumes:
      - application/json
      parameters:
      - description: properties to include, and optionally remap (comma separated)
        in: query
        name: props
        type: string
      - description: properties to group by (comma separated)
        in: query
        name: groupby
        type: string
      - description: properties to order by (comma separated, prefix with '~' to reverse)
        in: query
        name: order
        type: string
      - description: property value filter (a, !a, a&b, a|b, (a,b),  a%,  a%&!ab%)
        in: query
        items:
          type: string
        name: filters
        type: array
      - description: number of objets to include in response
        in: query
        name: limit
        type: integer
      - description: offset of the first objet to include in response
        in: query
        name: offset
        type: integer
      - description: turn off metadata in response
        in: query
        name: meta
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TableResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List tags
      tags:
      - tags
    post:
      consumes:
      - application/json
      description: Creating tags requires no privilege.
      parameters:
      - description: list of tags to create or update
        in: body
        name: tags
        required: true
        schema:
          items:
            $ref: '#/definitions/tables.Tag'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tables.Tag'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Create or update tags
      tags:
      - tags
...
  /tags/{id}/services:
    get:
      consumes:
      - application/json
      parameters:
      - description: properties to include, and optionally remap (comma separated)
        in: query
        name: props
        type: string
      - description: properties to group by (comma separated)
        in: query
        name: groupby
        type: string
      - description: properties to order by (comma separated, prefix with '~' to reverse)
        in: query
        name: order
        type: string
      - description: property value filter (a, !a, a&b, a|b, (a,b),  a%,  a%&!ab%)
        in: query
        items:
          type: string
        name: filters
        type: array
      - description: number of objets to include in response
        in: query
        name: limit
        type: integer
      - description: offset of the first objet to include in response
        in: query
        name: offset
        type: integer
      - description: turn off metadata in response
        in: query
        name: meta
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.TableResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List services having a specific tag
      tags:
      - tags
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"